update navbar footer
