+++ models/Form/UserForm.php
@@ -12,10 +12,10 @@ class UserForm extends Model
     public $id;
     public $username;
     public $password;
-//    public $firstName;
-//
-//    public $lastName;
-//    public $phoneNum;
+    //    public $firstName;
+    //
+    //    public $lastName;
+    //    public $phoneNum;
     public $role;
     public $profile_id;
 
@@ -24,46 +24,48 @@ class UserForm extends Model
         return [
             [['username', 'password'], 'required'],
             [['username'], 'string', 'max' => 255],
-            ['password', 'string', 'min' => 8, 'message'=>'Mật khẩu ít nhất 8 kí tự'],
+            ['password', 'string', 'min' => 8, 'message' => 'Mật khẩu ít nhất 8 kí tự'],
             ['password', 'match', 'pattern' => '/[!@#$%^&*(),.?":{}|<>]/', 'message' => 'Mật khẩu bao gồm ít nhất 1 kí tự đặc biệt.'],
-//            [['firstName', 'lastName'], 'string', 'max' => 255],
-//            [['phoneNum'], 'string', 'max' => 10],
+            //            [['firstName', 'lastName'], 'string', 'max' => 255],
+            //            [['phoneNum'], 'string', 'max' => 10],
             [['id', 'profile_id'], 'integer'],
             [['role'], 'string']
         ];
     }
 
-    public function save($user) {
-        if($this->validate()) {
-//            $user = new User();
+    public function save($user)
+    {
+        if ($this->role == '') {
+            $this->role = 'user';
+        }
+        if ($this->validate()) {
+            //            $user = new User();
             $user->username = $this->username;
             $user->password = Yii::$app->getSecurity()->generatePasswordHash($this->password);
-//            $user->created_at = time();
-//            $user->created_by = Yii::$app->user->identity->username;
+            //            $user->created_at = time();
+            //            $user->created_by = Yii::$app->user->identity->username;
             //$user->password = $this->password;
             $user->role = $this->role;
             $rs = $user->save();
             $auth = Yii::$app->authManager;
             $this->id = $user->id;
-//            return $rs;
+            //            return $rs;
             // var_dump($rs);
-            if($rs) {
-                if((UserProfile::findOne(['user_id'=>$this->id])) == null){
+            if ($rs) {
+                if ((UserProfile::findOne(['user_id' => $this->id])) == null) {
                     $user_profile = new UserProfile();
                     $user_profile->user_id = $this->id;
 
                     $pr  = $user_profile->save();
-//                     var_dump($pr);
-//                     die;
-                    if($pr) {
+                    //                     var_dump($pr);
+                    //                     die;
+                    if ($pr) {
                         $this->profile_id = $user_profile->id;
                         return $rs;
                     }
                 }
-
             }
         }
         return false;
     }
-
 }
