--- views/site/homepage.css
@@ -1,227 +1,0 @@
-*, *:before, *:after {
-  -webkit-box-sizing: border-box;
-  box-sizing: border-box;
-}
-
-body, .container, .full-width .ubermenu-nav, .container-width, .row {
-  width: 100%;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-.container {
-  /* padding-left: 15px;
-  padding-right: 15px; */
-}
-
-.container-width, .full-width .ubermenu-nav, .container, .row {
-  max-width: 1080px;
-}
-
-.section-title-container {
-  /* margin-bottom: 1.5em; */
-}
-
-/* @media (max-width: 549px) */
-h2 {
-  color: #000000;
-  font-weight: 700;
-  font-family: "Roboto", sans-serif;
-  font-size: 1.2em;
-}
-
-#header {
-  position: fixed;
-  width: 100%;
-  top: 0;
-  z-index: 1000; /* Set a high z-index to ensure it's above other elements */
-  background-color: #a7a6a6;
-  color: white;
-}
-
-#account a {
-    text-decoration: none;
-}
-
-#account a:hover {
-    margin-bottom: 2px;
-    border-bottom: 4px solid #585454;
-    
-}
-
-#account span {
-  font-weight: 500;
-}
-.divider{
-    border-left: 1px solid rgba(58, 55, 55, 0.1);
-    height: 30px;
-    vertical-align: middle;
-    position: relative;
-    /* margin: 0 7.5px; */
-}
-#navbarSupportedContent {
-  text-align: center;
-  background-color: #a5a3a3;
-  color: white ;
-  font-weight: 500;
-  font-size: medium;
-}
-
-
-#navbarSupportedContent .nav-item{
-  /* margin: 5px 15px; */
-}
-#navbarSupportedContent .dropdown-toggle{
-  background-color:black;
-  width: 100%;
-  /* padding: auto 5px; */
-}
-#navbarSupportedContent .nav-item .dropdown-menu .dropdown-item{
-  border-bottom: 1px solid #aaa;
-}
-#navbarSupportedContent .nav-item .dropdown-menu .dropdown-item:hover{
-  background-color: #a5a3a3;
-}
-
-#navbarSupportedContent ul .nav-link{
-  color: white ;
-}
-#navbarSupportedContent ul li.nav-item:not(.dropdown):hover a.nav-link {
-  color: black;
-}
-
-.navbar {
-  padding: 0;
-}
-
-footer{
-  font-family: 'Times New Roman', Times, serif;
-}
-footer .section .information{
-  background-color: black;
-  color: white;
-}
-
-footer .section-content .col {
-  padding: 10px 20px;
-}
-
-.flickity-page-dots .dot {
-  display: inline-block;
-  width: 12px;
-  height: 12px;
-  margin: 0 5px;
-  border: 3px solid #111;
-  border-radius: 50%;
-  cursor: pointer;
-  transition: opacity .3s;
-  /* position: absolute; */
-}
-
-.flickity-page-dots .dot .active {
-  background-color: #111111;
-  opacity: 1 !important;
-}
-
-.content {
-  box-sizing: border-box;
-  padding-top: 0;
-  /* margin-bottom: 40px; */
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  margin: 0 15% 40px 15%;
-}
-
-.slide {
-  margin-bottom: 40px;
-}
-
-.coffee {
-  /* padding: 20px 0; */
-  height: 70px;
-  position: relative;
-  /* padding: 20px; */
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 90%;
-}
-.box-line {
-  border: solid 1px #e5e5e5;
-  font-size: 20px;
-  position: absolute;
-  top: 15%;
-  padding: 10px;
-  background-color: white;
-
-}
-
-.line {
-  width: 100%;
-  height: 2px; 
-  background-color: #e5e5e5; 
-  /* margin: 0 auto; */
-  /* margin-top: 50px; */
-  position: absolute;
-  /* top: 40px; */
-  display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.col-text {
-  text-align: center;
-  margin-top: 15px;
-}
-
-.col-text p, .size {
-  display: inline-block;
-  font-size: small;
-}
-
-.width {
-  width: 100%;
-}
-
-.content-img {
-  width: 100%;
-  position: relative;
-  height: 70%;
-
-}
-
-.a {
-  border: none;
-}
-
-.container-divide {
-  display: flex;
-  width: 90%;
-  /* position: relative; */
-}
-
-.child {
-  flex: 1; /* Mỗi cột chiếm đủ không gian còn lại trong container */
-  margin-right: 20px; /* Khoảng cách giữa các cột */
-  width: 100%;
-}
-
-.border-radius {
-  border-radius: 100%;
-  width: 90%;
-}
-
-.rate {
-  display: flex;
-  /* flex-direction: column; */
-}
-
-.star1 {
-  flex: 1;
-  margin-right: 20px;
-}
-.star2 {
-  flex: 4;
-}
