+++ views/site/homepage.php
@@ -17,133 +17,9 @@
 </head>
 
 <body>
+
+    <?php include './navbar.php'; ?>
     <div class="content-container">
-        <header class="header has-sticky sticky-jump">
-            <div class="wrapper">
-                <div id="masthead" class="menu flex-row container-fluid row align-items-center logo-left medium-logo-center">
-                    <!-- Logo -->
-                    <div id="logo" class="col-3">
-                        <a href="homepage.php" title="Trung Nguyên E-Coffee" rel="home">
-                            <img width="200" height="100" alt="Trung Nguyên E-Coffee" src="https://trungnguyenecoffee.com/wp-content/uploads/2020/09/Logo-Màu-Trằng-E-Coffee.png" class="header_logo header-logo ls-is-cached lazyloaded" />
-                        </a>
-                    </div>
-                    <!-- Search Bar -->
-                    <div id="search-bar" class="col-4">
-                        <form role="search" method="get" class="search-form" action="https://trungnguyenecoffee.com/">
-                            <div class="input-group">
-                                <input type="search" id="woocommerce-product-search-field-0" class="form-control" placeholder="Tìm kiếm sản phẩm" value="" name="s" autocomplete="off" />
-                                <div class="input-group-append">
-                                    <button type="submit" value="Tìm kiếm" class="btn btn-secondary">
-                                        <i class="fas fa-search"></i>
-                                    </button>
-                                </div>
-                            </div>
-                            <div class="live-search-results text-left z-top">
-                                <div class="autocomplete-suggestions" style="
-                      position: absolute;
-                      display: none;
-                      max-height: 300px;
-                      z-index: 9999;
-                    "></div>
-                            </div>
-                        </form>
-                    </div>
-                    <!-- Signin/Logup -->
-                    <div id="account" class="col-5">
-                        <ul class="nav nav-right container-fluid d-flex align-items-center">
-                            <li class="col-6">
-                                <a href="register.php" title="" class="text-decoration-none text-dark">
-                                    <span>Đăng nhập / Đăng ký</span>
-                                </a>
-                            </li>
-                            <!-- <li class="col-2 divider"></li> -->
-                            <li class="col-4">
-                                <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-bag" viewBox="0 0 16 16">
-                                    <path d="M8 1a2.5 2.5 0 0 1 2.5 2.5V4h-5v-.5A2.5 2.5 0 0 1 8 1m3.5 3v-.5a3.5 3.5 0 1 0-7 0V4H1v10a2 2 0 0 0 2 2h10a2 2 0 0 0 2-2V4zM2 5h12v9a1 1 0 0 1-1 1H3a1 1 0 0 1-1-1z" />
-                                </svg>
-                                <a href="http://" class="text-decoration-none text-dark">
-                                    <span>Giỏ hàng</span>
-                                </a>
-                            </li>
-                        </ul>
-                    </div>
-                </div>
-                <nav class="navbar navbar-expand-lg">
-                    <div class="collapse navbar-collapse" id="navbarSupportedContent">
-                        <ul class="navbar-nav mx-auto mb-2 mb-lg-0">
-                            <!-- Danh mục sản phẩm (Dropdown menu) -->
-                            <li class="nav-item dropdown">
-                                <a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
-                                    <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-list" viewBox="0 0 16 16">
-                                        <path fill-rule="evenodd" d="M2.5 12a.5.5 0 0 1 .5-.5h10a.5.5 0 0 1 0 1H3a.5.5 0 0 1-.5-.5m0-4a.5.5 0 0 1 .5-.5h10a.5.5 0 0 1 0 1H3a.5.5 0 0 1-.5-.5m0-4a.5.5 0 0 1 .5-.5h10a.5.5 0 0 1 0 1H3a.5.5 0 0 1-.5-.5" />
-                                    </svg>
-                                    <span>Danh mục sản phẩm</span>
-                                </a>
-                                <div class="dropdown-menu" aria-labelledby="dropdownMenuButton">
-                                    <a class="dropdown-item" href="#/Caphedonggoi">
-                                        <img width="30" height="30" alt="cà phê đóng gói" decoding="async" loading="lazy" src="https://trungnguyenecoffee.com/wp-content/uploads/2021/08/Icon-web_Ca-phe-dong-goi.png" />
-                                        <span>Cà phê đóng gói</span>
-                                    </a>
-                                    <a class="dropdown-item" href="#/Quatangcaocap">
-                                        <img width="30" height="30" alt="quà tặng cao cấp" decoding="async" loading="lazy" src="https://trungnguyenecoffee.com/wp-content/uploads/2021/12/Icon-web-gift.png" />
-                                        <span>Quà tặng cao cấp</span>
-                                    </a>
-                                    <a class="dropdown-item" href="#/Vatphambanle">
-                                        <img width="30" height="30" alt="vật phẩm bán lẻ" decoding="async" loading="lazy" src="https://trungnguyenecoffee.com/wp-content/uploads/2021/08/Icon-web_Vat-pham-ca-phe.png" />
-                                        <span>Vật phẩm bán lẻ</span>
-                                    </a>
-                                    <a class="dropdown-item" href="#/Vatphamsachquy">
-                                        <img width="30" height="30" alt="vật phẩm sách quý" decoding="async" loading="lazy" src="https://trungnguyenecoffee.com/wp-content/uploads/2021/08/Icon-web_San-pham-sach-quy.png" />
-                                        <span>Vật phẩm sách quý</span>
-                                    </a>
-                                    <a class="dropdown-item" href="#/Vatphamdoitac">
-                                        <img width="30" height="30" alt="vật phẩm đối tác" decoding="async" loading="lazy" src="https://trungnguyenecoffee.com/wp-content/uploads/2021/08/Icon-web_San-pham-doi-tac.png" />
-                                        <span>Vật phẩm đối tác</span>
-                                    </a>
-                                    <a class="dropdown-item" href="#/maymocthietbi">
-                                        <img width="30" height="30" alt="máy móc thiết bị" decoding="async" loading="lazy" src="https://trungnguyenecoffee.com/wp-content/uploads/2021/08/Icon-web_May-moc-thiet-bi.png" />
-                                        <span> Máy móc thiết bị</span>
-                                    </a>
-                                    <a class="dropdown-item" href="#/Congcudungcu">
-                                        <img width="30" height="30" alt="cà phê đóng gói" decoding="async" loading="lazy" src="https://trungnguyenecoffee.com/wp-content/uploads/2021/08/Icon-web_Cong-cu-dung-cu.png" />
-                                        <span>Công cụ dụng cụ</span>
-                                    </a>
-                                    <a class="dropdown-item" href="#/Baobithuonghieu">
-                                        <img width="30" height="30" alt="cà phê đóng gói" decoding="async" loading="lazy" src="https://trungnguyenecoffee.com/wp-content/uploads/2021/08/Icon-web_Bao-bi-thuong-hieu.png" />
-                                        <span>Bao bì thương hiệu</span>
-                                    </a>
-                                    <a class="dropdown-item" href="#">
-                                        <img width="30" height="30" alt="cà phê đóng gói" decoding="async" loading="lazy" src="https://trungnguyenecoffee.com/wp-content/uploads/2021/08/Icon-web_Ca-phe-dong-goi.png" />
-                                        Cà phê đóng gói
-                                    </a>
-                                </div>
-                            </li>
-                            <li class="nav-item">
-                                <a class="nav-link" href="#">Câu chuyện thương hiệu</a>
-                            </li>
-                            <li class="nav-item">
-                                <a class="nav-link" href="#">Nhượng quyền</a>
-                            </li>
-                            <li class="nav-item">
-                                <a class="nav-link" href="#">Sản phẩm</a>
-                            </li>
-                            <li class="nav-item">
-                                <a class="nav-link" href="#">Khuyến mãi</a>
-                            </li>
-                            <li class="nav-item">
-                                <a class="nav-link" href="#">Tin tức</a>
-                            </li>
-                            <li class="nav-item">
-                                <a class="nav-link" href="#">Cửa hàng</a>
-                            </li>
-                            <li class="nav-item">
-                                <a class="nav-link" href="#">Tuyển dụng</a>
-                            </li>
-                        </ul>
-                    </div>
-                </nav>
-            </div>
-        </header>
         <div id="carouselExampleIndicators" class="carousel slide" data-ride="carousel">
             <ol class="carousel-indicators flickity-page-dots">
                 <li data-target="#carouselExampleIndicators" data-slide-to="0" class="active dot" aria-label="Page dot 1"></li>
@@ -205,29 +81,6 @@
                     <strong>Thông Tin Nhượng Quyền Trung Nguyên E-Coffee</strong>
                 </div>
             </div>
-            <!-- <div class="coffee">
-                <div class="col col-border">
-                    <div class="line">
-                        <div class="box-line">
-                            <strong>GÓI THỊNH VƯỢNG</strong>
-                        </div>
-                    </div>
-                </div>
-                <div class="col col-border">
-                    <div class="line">
-                        <div class="box-line">
-                            <strong>GÓI KHỞI NGHIỆP</strong>
-                        </div>
-                    </div>
-                </div>
-                <div class="col col-border">
-                    <div class="line">
-                        <div class="box-line">
-                            <strong>GÓI KẾT NỐI</strong>
-                        </div>
-                    </div>
-                </div>
-            </div> -->
             <div class="container-divide">
                 <div class="child">
                     <div class="coffee width">
@@ -353,56 +206,13 @@
             </div>
 
         </div>
-        <footer id="footer" class="footer-wrapper">
-            <section class="section ">
-                <div class="information">
-                    <div class="section-content">
-                        <div class="row row-collapse row-ful-width hide-for-medium">
-                            <div class="col text-left pl-1">
-                                <h4 class="uppercase">
-                                    CÔNG TY CỔ PHẦN TRUNG NGUYÊN FRANCHISING
-                                </h4>
-                                <p>
-                                    87A Cách Mạng Tháng 8, Phường Bến Thành, Quận 1, Hồ Chí Minh
-                                </p>
-                                <p>Hotline: 1900 96 96 68</p>
-                                <p>Liên hệ hợp tác Nhượng Quyền E-Coffee:</p>
-                                <p>Hotline1: 088 673 1188</p>
-                                <p>Hotline2: 081 673 1188</p>
-                                <p>GPKD: 0310939343</p>
-                            </div>
-                            <div class="col text-left">
-                                <h4 class="uppercase">LIÊN KẾT NHANH</h4>
-                                <p>Trung Nguyên Legend Café</p>
-                                <p>Thế Giới Cà Phê</p>
-                                <p>Trung Nguyên Legend</p>
-                                <p>Bảo Tàng Thế Giới Cà Phê</p>
-                                <p>Tra cứu hóa đơn điện tử</p>
-                                <p>Đóng góp ý kiến</p>
-                            </div>
-                            <div class="col text-left">
-                                <h4 class="uppercase">HỖ TRỢ KHÁCH HÀNG</h4>
-                                <p>Chính sách bảo mật</p>
-                                <p>Chính sách và điều khoản</p>
-                                <p>Chính sách giao nhận hàng</p>
-                                <p>Chính sách bảo hành và đổi trả</p>
-                                <p>Chính sách hoàn trả hàng và hoàn tiền</p>
-                            </div>
-                            <div class="col text-left">
-                                <h4 class="uppercase">FANPAGE</h4>
-                                <p>Trung Nguyên E-Coffee</p>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </section>
-        </footer>
-
-        <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous">
-        </script>
-        <script src="https://cdn.jsdelivr.net/npm/popper.js@1.12.9/dist/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous">
-        </script>
-        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"></script>
+    </div>
+    <?php include './footer.php'; ?>
+    <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous">
+    </script>
+    <script src="https://cdn.jsdelivr.net/npm/popper.js@1.12.9/dist/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous">
+    </script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"></script>
 
 
 </body>
