--- views/site/reset_password.css
@@ -1,80 +1,0 @@
-#header {
-    position: fixed;
-    width: 100%;
-    top: 0;
-    z-index: 1000;
-    background-color: #a7a6a6;
-    color: white;
-}
-#sidebar-toggle{
-    border: none;
-}
-  
-#navbarSupportedContent {
-    text-align: center;
-    background-color: #a5a3a3;
-    color: white ;
-    font-weight: 500;
-    font-size: medium;
-}
-#navbarSupportedContent .nav-item{
-    margin: 5px 15px;
-}
-#navbarSupportedContent .dropdown-toggle{
-    background-color:black;
-    width: 100%;
-    padding: auto 5px;
-}
-#navbarSupportedContent .nav-item .dropdown-menu .dropdown-item{
-    border-bottom: 1px solid #aaa;
-}
-#navbarSupportedContent .nav-item .dropdown-menu .dropdown-item:hover{
-    background-color: #a5a3a3;
-}
-  
-#navbarSupportedContent ul .nav-link{
-    color: white ;
-}
-#navbarSupportedContent ul li.nav-item:not(.dropdown):hover a.nav-link {
-    color: black;
-}
-#account a {
-    text-decoration: none;
-}
-  
-#account a:hover {
-    margin-bottom: 2px;
-    border-bottom: 4px solid #585454;
-    
-}
-  
-#account span {
-  font-weight: 500;
-}
-#reset-password{
-    margin: 200px ;
-}
-#reset-password .form-group label{
-    margin-bottom: 15px;
-    font-weight: bold;
-}
-#reset-password .btn {
-    background-color: black;
-    color: white;
-    font-weight: bold;
-    border-radius: 0;
-    margin-top: 15px;
-    padding: 10px 15px;
-}
-
-
-footer{
-    font-family: 'Times New Roman', Times, serif;
-}
-footer .information{
-    background-color: black;
-}
-footer .section-content .col {
-    padding: 10px 20px;
-}
-
