+++ web/css/homepage.css
@@ -1,0 +1,124 @@
+.container, html {
+  padding: 0;
+  margin: 0;
+}
+
+.flickity-page-dots .dot {
+  display: inline-block;
+  width: 12px;
+  height: 12px;
+  margin: 0 5px;
+  border: 3px solid #111;
+  border-radius: 50%;
+  cursor: pointer;
+  transition: opacity .3s;
+  /* position: absolute; */
+}
+
+.flickity-page-dots .dot .active {
+  background-color: #111111;
+  opacity: 1 !important;
+}
+
+.content {
+  box-sizing: border-box;
+  padding-top: 0;
+  /* margin-bottom: 40px; */
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  margin: 0 15% 40px 15%;
+}
+
+.slide {
+  margin-bottom: 40px;
+}
+
+.coffee {
+  /* padding: 20px 0; */
+  height: 70px;
+  position: relative;
+  /* padding: 20px; */
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 90%;
+}
+.box-line {
+  border: solid 1px #e5e5e5;
+  font-size: 20px;
+  position: absolute;
+  top: 15%;
+  padding: 10px;
+  background-color: white;
+
+}
+
+.line {
+  width: 100%;
+  height: 2px; 
+  background-color: #e5e5e5; 
+  /* margin: 0 auto; */
+  /* margin-top: 50px; */
+  position: absolute;
+  /* top: 40px; */
+  display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.col-text {
+  text-align: center;
+  margin-top: 15px;
+}
+
+.col-text p, .size {
+  display: inline-block;
+  font-size: small;
+}
+
+.width {
+  width: 100%;
+}
+
+.content-img {
+  width: 100%;
+  position: relative;
+  height: 70%;
+
+}
+
+.a {
+  border: none;
+}
+
+.container-divide {
+  display: flex;
+  width: 90%;
+  /* position: relative; */
+}
+
+.child {
+  flex: 1; /* Mỗi cột chiếm đủ không gian còn lại trong container */
+  margin-right: 20px; /* Khoảng cách giữa các cột */
+  width: 100%;
+}
+
+.border-radius {
+  border-radius: 100%;
+  width: 90%;
+}
+
+.rate {
+  display: flex;
+  /* flex-direction: column; */
+}
+
+.star1 {
+  flex: 1;
+  margin-right: 20px;
+}
+.star2 {
+  flex: 4;
+}
