+++ web/css/navbar.css
@@ -1,0 +1,92 @@
+*, *:before, *:after {
+  -webkit-box-sizing: border-box;
+  box-sizing: border-box;
+}
+
+body, .container, .full-width .ubermenu-nav, .container-width {
+  width: 100%;
+}
+
+
+.container-width, .full-width .ubermenu-nav, .container{
+  max-width: 100%;
+}
+
+h2 {
+  color: #000000;
+  font-weight: 700;
+  font-family: "Roboto", sans-serif;
+  font-size: 1.2em;
+}
+
+#header {
+  position: fixed;
+  width: 100%;
+  top: 0;
+  z-index: 1000; /* Set a high z-index to ensure it's above other elements */
+  background-color: #a7a6a6;
+  color: white;
+}
+
+#account a {
+    text-decoration: none;
+}
+
+#account a:hover {
+    margin-bottom: 2px;
+    border-bottom: 4px solid #585454;
+    
+}
+
+#account span {
+  font-weight: 500;
+}
+.divider{
+    border-left: 1px solid rgba(58, 55, 55, 0.1);
+    height: 30px;
+    vertical-align: middle;
+    position: relative;
+    /* margin: 0 7.5px; */
+}
+#navbarSupportedContent {
+  text-align: center;
+  background-color: #a5a3a3;
+  color: white ;
+  font-weight: 500;
+  font-size: medium;
+}
+
+
+#navbarSupportedContent .dropdown-toggle{
+  background-color:black;
+  width: 100%;
+}
+#navbarSupportedContent .nav-item .dropdown-menu .dropdown-item{
+  border-bottom: 1px solid #aaa;
+}
+#navbarSupportedContent .nav-item .dropdown-menu .dropdown-item:hover{
+  background-color: #a5a3a3;
+}
+
+#navbarSupportedContent ul .nav-link{
+  color: white ;
+}
+#navbarSupportedContent ul li.nav-item:not(.dropdown):hover a.nav-link {
+  color: black;
+}
+
+.navbar.navbar-expand-lg {
+  padding: 0;
+}
+
+.row.align-items-center {
+  margin: 0 10%;
+}
+/* 
+.wrapper {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: row;
+} */
+
