+++ web/css/register.css
@@ -1,0 +1,38 @@
+#register {
+    font-family: Arial, sans-serif;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    height: 100vh;
+}
+
+h4 {
+    margin-bottom: 40px;
+    text-transform: uppercase;
+    font-weight: bold;
+}
+
+.divider {
+    border-right: 1px solid #aaa;
+    height: 100%;
+}
+#register .form-group label,
+#register .form-group input[type="text"] {
+    font-weight: bold;
+    margin-bottom: 15px;
+}
+
+#register .btn {
+    background-color: black;
+    color: white;
+    font-weight: bold;
+    border-radius: 0;
+    padding: 10px 15px;
+
+}
+
+#register a {
+    color: black;
+    text-decoration: none;
+}
+
